moved APITest, SampleAPI test classes to 'internal.deployment' test package
